Versiune Adi/nereus-system-main/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.18
pydantic-settings>=2.3.0
numpy>=1.26.0
scipy>=1.13.0
rasterio>=1.3.10
openeo>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
aiofiles>=23.2.1
scikit-learn>=1.3.0
joblib>=1.3.0
shapely>=2.0.0
resend>=0.7.0
